Change LeaderboardRatingJournalBean to immutable

+++ src/main/java/com/faforever/client/domain/GamePlayerStatsBean.java
@@ -6,9 +6,7 @@ import java.time.OffsetDateTime;
 import java.util.List;
 
 public record GamePlayerStatsBean(
-    PlayerBean player,
-    int score,
-    int team,
+    PlayerBean player, byte score, byte team,
     Faction faction,
     OffsetDateTime scoreTime,
     ReplayBean game,
